intermediate..
We should decode NAS message, and then integrity is checked later.
Because, To get MME UE Context, at first, we need to know GUTI in NAS
message.
if S-TMSI is not available.. Sigh!

if NAS is ciphered, The seqence are 'Decrypt->Decode->Integrity Check'
if NAS is not cihpered, The sequence are 'Decode->Integrity Check'