[SMF] Store MSISDN from GTPC and pass it in Gy CCR
The encoding in GTPv1C is different than in GTPv2C. The former has an
extra byte prepended specifying the nature of address and numbering plan
indicator. Diameter doesn't require that information so we simply strip
it of.
In any case TS 29.060 (GTPv1C) sec 7.7.33 states:
"""
The MSISDN shall be in international format and the "nature of address indicator" shall indicate
"international number".
"""
So we are fine stripping and copying over.