[SMF] Gn: Only handle and report back APN-AMBR IE if requested by SGSN
As stated in 3GPP TS 29.060 7.3.2
"""
The APN-AMBR IE shall be included as the authorized APN-AMBR if the GGSN supports
this IE and if the APN-AMBR IE has been included in the corresponding request message.
"""

Take the chance also to initialize SMF AMBR values if this IE is
received, since it can be that QoS Profile IE doesn't contain this
information.