[SMF] Gn: Round APN-AMBR values upwards when converting to kbps
As stated in TS 29.060 7.7.98:
"""
The APN-AMBR for uplink and the APN-AMBR for downlink may require converting values in bits per second to
kilobits per second when the APN-AMBR for uplink and the APN-AMBR for downlink are received from an interface
other than GTP interface. If such conversions result in fractions, then the value of APN-AMBR for uplink and the APN-
AMBR for downlink shall be rounded upwards.
"""