[metrics] Fix double-free on application exit
Adding metric instance to the list used "&inst->entry", while trying to
remove it from the list used pointer "inst". This has caused that the
instance was freed, but the pointer still remained in the list.

On application exit, each application cleanups it's own metrics, while
metrics library does the same. This has caused double-free crash.