[MME] Set TAI list of InitialContextSetupRequest in S1AP/NAS-EPS for UE/RAN of OAI.
When using OAI for UE/RAN (use_openair == true), set TAI list of InitialContextSetupRequest to TAI (PLMN x 1, TAC x 1). The TAI information at this time is used by copying the mme_ue->tai information as is. The reason is that OAI can parse the TAI list of InitialContextSetupRequest in S1AP/NAS-EPS only in this format, but fails to parse the TAI list in other formats, such as the following.

- (PLMN x 1, TAC x n)
- (PLMN x 1, TAC x 1) x n

In mme section of mme.yaml

mme:
    ...
    tai:
      plmn_id:
        mcc: 999
        mnc: 70
      tac: [1, 2]

or

mme:
    ...
    tai:
      - plmn_id:
          mcc: 999
          mnc: 70
        tac: 1
      - plmn_id:
          mcc: 999
          mnc: 70
        tac: 2

If configured as above, TAI list in InitialContextSetupRequest in S1AP/NAS-EPS will be sent to UE/RAN with values in such format, but OAI will fail to process this format data.

When trying functions of Open5GS CUPS-EPC, it is very convenient to use OAI's UE/RAN simulator and so I will pull request this.