[AMF] Replace assert() with expect() when sending NGAP and SBI messages
It is possible that AMF receives a response to some SBI message request,
which requires an NGAP message to be sent, but UE or even gNB could be
disconnected already in the meantime.
This change prevents AMF to crash on assert if ran_ue or gnb contexts
are already deallocated.

It is possible that a message from gNB arrives at AMF before AMF
successfully registers to the NRF at application startup. In that case,
AMF crashes beacuse it can't discover other NF services (for example
AUSF).