[NF-DISCOVERY] When resolving fqdn, try to use port from ipEndPoints
TS 29.510 - 5G System- Network function repository services, chapter 6.2.6.2.4
- Type: NFService (148), NOTE 5:

If the NF Service Consumer, based on the FQDN and IP address related attributes
of the NFProfile and NFService, determines that it needs to use an FQDN to
establish the HTTP connection with the NF Service Producer, it shall use such
FQDN for DNS query and, in absence of any port information in the ipEndPoints
attribute of the NF Service, it shall use the default HTTP port number, i.e.
TCP port 80 for "http" URIs or TCP port 443 for "https" URIs as specified in
IETF RFC 7540 [9] when invoking the service.