[SMF] Do not check policy in case of DELETE_TRIGGER_UE_REQUESTED
It is the case in which the policy had already been terminated.
Since SMF did not remove policy information after termination,
this bug did not show up.

Bug:

After proper removing of policy information, the Establishment
was rejected with Error Indication in ue-context-test.

Fix:

Policy information is not checked in case of
SMF_NGAP_STATE_DELETE_TRIGGER_UE_REQUESTED.