[SBI] Fix finding NF service by it's ID
The issue was that the function returned the last NF service of the
list, even if the ID was not found in the list.
The fixed version is in-line with the function
ogs_sbi_nf_service_find_by_name().

The function is used in parsing the response from NRF, compiling an
internal list of NF's and it's services. The bug could cause that some
services of NF would not be represented.