[MME] SGsAP TMSI-REALLOCATION-COMPLETE (#3381)
TS24.301
5.5.3 Tracking area updating procedure (S1 mode only)
5.5.3.2 Normal and periodic tracking area updating procedure
5.5.3.2.4 Normal and periodic tracking area updating procedure
accepted by the network

According to the standard, the MME should verify whether
the P-TMSI has changed. If it has changed, the MME must
wait for the Tracking Area Update Complete message. After
sending the Tracking Area Update Accept, the T3450 timer
should be started to wait for the Complete message. During
this time, the EMM state should remain in
initial_context_setup_request state. Only upon receiving the
Complete message should the state transition to
EMM-Registered.

However, for simplicity, this implementation in Open5GS
MME does not verify whether the P-TMSI has changed.
Additionally, the T3450 timer is not started, and the
state is immediately transitioned to EMM-Registered.

If the UE sends the Tracking Area Update Complete, it
implies that the P-TMSI has changed. Therefore, when
receiving the Tracking Area Update Complete message in the
EMM-Registered state, the implementation sends an SGsAP
TMSI-REALLOCATION-COMPLETE to the VLR.