[SBI] Support responses larger than MAX_FRAME_SIZE
Bug:

If response is larger than 16384 characters, it is not sent
by HTTP/2 server.

Fix:

nghttp2 supports sending of responses in chunks.
Now nghttp2 is used the way so that large responses are sent in chunks.
An offset is introduced to track how much of data has already
been sent over the HTTP/2 connection, so we know where to resume
at next callback.
Both variants are supported - USE_SEND_DATA_WITH_NO_COPY 0 and 1.
A zero-copy variant uses additional `on_send_data()` callback,
so an additional send_offset is introduced.