[AMF] Rej. sess with cause 65 if max. no. of PDU sess. is reached
If PDU_SESSION_ESTABLISHMENT_REQUEST is rejected, the AMF returns the
5GSM message which was not forwarded instead of 5GSM Status.

In case the maximum number of PDU sessions has already been reached for
the UE, the AMF shall send back to the UE 5GSM message with cause 65.

TS 24.501:

5.4.5.3 Network-initiated NAS transport procedure

5.4.5.3.1 General
The purpose of the network-initiated NAS transport procedure is to provide
a transport of:

h) a single uplink 5GSM message which was not forwarded, because the PLMN's
maximum number of PDU sessions has been reached;

5.4.5.3.2 Network-initiated NAS transport procedure initiation
In case h) in subclause 5.4.5.3.1, i.e. upon sending a single uplink 5GSM message
which was not forwarded, because the PLMN's maximum number of PDU sessions
has been reached, the AMF shall:

a) include the PDU session ID in the PDU session ID IE;
b) set the Payload container type IE to "N1 SM information";
c) set the Payload container IE to the 5GSM message which was not forwarded; and
d) set the 5GMM cause IE to the 5GMM cause #65 "maximum number of PDU sessions
   reached".