[AMF] Fix crash in gmm_state_de_registered due to invalid state handling (#4082)
Problem
-------
AMF was crashing several times per day when multiple UEs were connecting,
sending data, and disconnecting. The crash occurred in
gmm_state_de_registered() with the following log:

    FATAL: Invalid state [2]
    FATAL: gmm_state_de_registered: should not be reached.

This was triggered by network-initiated deregistration scenarios that were
not explicitly handled in the FSM, causing the AMF to enter an unexpected
state and abort.