[AMF] fix checking for correct serving TAI
In case gNB sends a NGSetupRequest with a TAI that AMF does not serve,
AMF would in some cases respond with a NGSetupSuccess instead of
NGSetupFailure.
This is because of a missing check for correct PLMN.

For example, AMF was configured with the following:
PLMN 001-01:
    tac: [1000-1100]
PLMN 999-93:
    tac: [2000-2200]

gNB sent NGSetupRequest for PLMN 999-93 and with TAC 1050. AMF would
only check the TAC, instead of also PLMN.